requirements.txt
Pillow
exifread
tk